reorg, docs, bump to 0.2.0
