+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "rdkit-sys"
 authors = ["Xavier Lange (xrlange@gmail.com)", "chrissly31415"]
-version = "0.1.12"
+version = "0.2.0"
 edition = "2021"
 license = "MIT"
 description = "RDKit CFFI library builder and bindings"
@@ -12,10 +12,8 @@ exclude = ["rdkit-*", "*.tar.gz"]
 
 [dependencies]
 cxx = "1.0.66"
-libc = "0.2"
 
 [build-dependencies]
 pkg-config = "0.3.9"
 env_logger = "0"
-eyre = "0.6"
 cxx-build = "1.0.67"
