+++ src/bridge/fingerprint.rs
@@ -1,9 +1,13 @@
 #[cxx::bridge(namespace = "RDKit")]
 pub mod ffi {
     unsafe extern "C++" {
-        include!("wrapper/include/rdmol.h");
+        include!("wrapper/include/ro_mol.h");
+        include!("wrapper/include/fingerprint.h");
 
+        pub type ROMol = crate::ro_mol::ffi::ROMol;
         pub type ExplicitBitVect;
+        pub fn fingerprint_mol(mol: SharedPtr<ROMol>) -> SharedPtr<ExplicitBitVect>;
+
         pub fn copy_explicit_bit_vect(
             fingerprint: SharedPtr<ExplicitBitVect>,
         ) -> SharedPtr<ExplicitBitVect>;
