+++ src/bridge/mod.rs
@@ -1,3 +1,4 @@
 pub mod fingerprint;
 pub mod mol_standardize;
-pub mod rdmol;
+pub mod ro_mol;
+pub mod rw_mol;
