+++ src/bridge/ro_mol.rs
@@ -1,13 +1,12 @@
 #[cxx::bridge(namespace = "RDKit")]
 pub mod ffi {
     unsafe extern "C++" {
-        include!("wrapper/include/rdmol.h");
+        include!("wrapper/include/ro_mol.h");
 
         pub type ROMol;
-        pub type ExplicitBitVect = crate::cxx::bridge::fingerprint::ffi::ExplicitBitVect;
+        pub type ExplicitBitVect = crate::fingerprint::ffi::ExplicitBitVect;
 
         pub fn mol_from_smiles(smi: &CxxString) -> SharedPtr<ROMol>;
         pub fn mol_to_smiles(mol: SharedPtr<ROMol>) -> String;
-        pub fn fingerprint_mol(mol: SharedPtr<ROMol>) -> SharedPtr<ExplicitBitVect>;
     }
 }
