+++ wrapper/src/rw_mol.cc
@@ -1,0 +1,9 @@
+#include "rust/cxx.h"
+#include "GraphMol/GraphMol.h"
+
+namespace RDKit {
+    std::shared_ptr<RWMol> rw_mol_from_ro_mol(std::shared_ptr<ROMol> mol, bool quick_copy, int conf_id) {
+        RWMol *rw_mol = new RWMol(*mol, quick_copy, conf_id);
+        return std::shared_ptr<RWMol>(rw_mol);
+    }
+}
